Use a Scope api, not a context api
